+++ samples/net/Demo.ActivateNonModalDialog/ViewModelLocator.cs
@@ -1,0 +1,22 @@
+using CommonServiceLocator;
+using GalaSoft.MvvmLight.Ioc;
+
+namespace Demo.ActivateNonModalDialog
+{
+    /// <summary>
+    /// This class contains static references to all the view models in the
+    /// application and provides an entry point for the bindings.
+    /// </summary>
+    public class ViewModelLocator
+    {
+        public ViewModelLocator()
+        {
+            ServiceLocator.SetLocatorProvider(() => SimpleIoc.Default);
+
+            SimpleIoc.Default.Register<MainWindowViewModel>();
+            SimpleIoc.Default.Register<CurrentTimeDialogViewModel>();
+        }
+
+        public MainWindowViewModel MainWindow => ServiceLocator.Current.GetInstance<MainWindowViewModel>();
+    }
+}